# src/openapi.yaml
openapi: 3.1.0
info:
  title: MCP REST Bridge
  version: 1.0.0
  description: >
    A self-hosted MCP gateway that bridges MCP clients to backend REST APIs and
    other MCP servers, bidirectionally. REST-capable services register
    themselves (via OpenAPI/Swagger auto-discovery, a cURL/Postman import, or a
    manual tool list) and the bridge exposes their tools to any MCP client
    (e.g. Claude Desktop, Cursor, or a custom agent) via the Streamable HTTP
    MCP transport; existing MCP servers can likewise be registered as
    upstreams and re-exposed. Every call is proxied through a uniform guard
    pipeline (SSRF check, guardrails, per-tool policy, circuit breaker,
    response sanitizing, audit) documented in this admin API — this
    specification covers that management surface, not the dynamic per-client
    tool endpoints it provisions.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
...
paths:
...
  /admin-api/bundles:
    get:
      operationId: listBundles
      summary: List admin-curated tool bundles
      tags: [Bundles]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      responses:
        "200":
          description: All bundles.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      properties: { items: { type: array, items: { $ref: "#/components/schemas/BundleSummary" } } },
                    },
                },
            }
    post:
      operationId: createBundle
      summary: Create a new bundle
      description: A bundle is served on its own MCP endpoint at /mcp-custom/{name} once created. Requires a super-admin session (admin role, no team) — team-scoped admins are rejected.
      tags: [Bundles]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, pattern: "^[a-z0-9][a-z0-9_-]{0,62}$" }
                description: { type: string }
                tools:
                  {
                    type: array,
                    items: { $ref: "#/components/schemas/BundleToolRef" },
                    description: "Each entry {client, tool} must reference an existing, resolvable tool.",
                  }
                composites:
                  {
                    type: array,
                    items: { type: string },
                    description: "Names of existing composite (macro) tools this bundle also exposes.",
                  }
      responses:
        "201":
          description: Created.
          content: { application/json: { schema: { $ref: "#/components/schemas/BundleDetail" } } }
        "400":
          description: Invalid name, tools[]/composites[] shape, or an entry references an unknown client/tool/composite.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "403":
          description: Caller is not a super-admin.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "409":
          description: A bundle with that name already exists.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin-api/bundles/{name}:
    get:
      operationId: getBundle
      summary: Get full detail for one bundle
      tags: [Bundles]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Bundle detail.
          content: { application/json: { schema: { $ref: "#/components/schemas/BundleDetail" } } }
        "404":
          description: No such bundle.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    patch:
      operationId: updateBundle
      summary: Update a bundle's description, enabled state, tool list, and/or composite list
      description: Requires a super-admin session (admin role, no team) — team-scoped admins are rejected.
      tags: [Bundles]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                description: { type: [string, "null"] }
                enabled: { type: boolean }
                tools: { type: array, items: { $ref: "#/components/schemas/BundleToolRef" } }
                composites: { type: array, items: { type: string } }
      responses:
        "200":
          description: Updated.
          content:
            {
              application/json:
                { schema: { type: object, properties: { status: { type: string }, name: { type: string } } } },
            }
        "400":
          description: Invalid tools[]/composites[] shape, or an entry references an unknown client/tool/composite.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "403":
          description: Caller is not a super-admin.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: No such bundle.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    delete:
      operationId: deleteBundle
      summary: Delete a bundle
      description: Requires the admin role for session callers.
      tags: [Bundles]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Deleted.
          content:
            {
              application/json:
                {
                  schema:
                    {
                      type: object,
                      properties: { status: { type: string, example: deleted }, name: { type: string } },
                    },
                },
            }
        "404":
          description: No such bundle.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin-api/bundles/{name}/install-links:
    get:
      operationId: listInstallLinks
      summary: List a bundle's shareable install links
      description: The one-time tokens are never included — only link metadata (id, created/expiry, revocation state).
      tags: [Bundles]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Install links for this bundle.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: "#/components/schemas/InstallLink" } }
        "404":
          description: No such bundle.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
    post:
      operationId: createInstallLink
      summary: Mint a shareable "install this bundle" link
      description: >
        Creates a revocable install link backed by a fresh MCP API key scoped
        only to this bundle's tools. The raw token is returned exactly once, in
        this response — it is never persisted in cleartext or retrievable again.
        Requires the admin role for session callers.
      tags: [Bundles]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expiresAt:
                  type: [integer, "null"]
                  description: Unix ms epoch after which the link stops working; null/omitted means no expiry.
      responses:
        "201":
          description: Link created. `token` is present here and nowhere else.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/InstallLink"
                  - type: object
                    properties:
                      token: { type: string, description: The one-time raw token — shown only in this response. }
        "400":
          description: Invalid expiresAt.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
        "404":
          description: No such bundle.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
  /admin-api/bundles/{name}/install-links/{id}:
    delete:
      operationId: revokeInstallLink
      summary: Revoke an install link
      description: >
        Immediately invalidates the link and its scoped MCP key. Requires the
        admin role for session callers.
      tags: [Bundles]
      security: [{ AdminAuth: [] }, { SessionAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Revoked.
          content:
            {
              application/json:
                {
                  schema:
                    { type: object, properties: { status: { type: string, example: revoked }, id: { type: integer } } },
                },
            }
        "404":
          description: No such install link for this bundle.
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiError" } } }
...
components:
...
  securitySchemes:
    AdminAuth:
      type: http
      scheme: bearer
      description: >
        API key for administrative endpoints (registration, introspection).
        Set via ADMIN_API_KEYS env var. Disabled when AUTH_DISABLED=true.
    McpAuth:
      type: http
      scheme: bearer
      description: >
        API key for the backend-tool MCP transport endpoints
        (/mcp/{clientName}, /mcp-custom/{bundleName}). Set via MCP_API_KEYS
        env var, or a managed key (see MCP Keys). Disabled when
        AUTH_DISABLED=true or when no keys are configured at all.
    RootMcpAuth:
      type: http
      scheme: bearer
      description: >
        Credential for the /mcp system control-plane. Either the env admin
        Bearer key (ADMIN_API_KEYS) or a managed MCP key with `adminRole`
        set (see MCP Keys). Unlike McpAuth, there is no "no auth material
        configured => allow all" fallback — a caller must resolve to a real
        system role or the request is rejected outright, even on a fresh
        install with zero keys minted. Still bypassed entirely when
        AUTH_DISABLED=true (the same global dev/test escape hatch every
        other auth path in this API respects).
    SessionAuth:
      type: apiKey
      in: cookie
      name: __Host-mcp_admin_session
      description: >
        Admin-UI session cookie, issued by POST /admin-api/auth/login.
        Accepted anywhere AdminAuth is accepted. On mutating requests
        (anything but GET/HEAD/OPTIONS) an `X-CSRF-Token` header matching the
        `__Host-mcp_admin_csrf` cookie value is also required — OpenAPI 3.1
        has no first-class way to express that pairing, so it's documented
        here in prose, mirroring how RegistrationPayload's tools/openapi_url
        mutual-exclusivity is documented rather than schema-encoded.
        The `__Host-` prefix (and therefore this exact name) only applies
        when SESSION_COOKIE_SECURE=true (the default) — it mandates the
...
  schemas:
...
    BundleToolRef:
      type: object
      required: [client, tool]
      properties:
        client: { type: string }
        tool: { type: string }
    BundleSummary:
      type: object
      properties:
        name: { type: string }
        description: { type: [string, "null"] }
        enabled: { type: boolean }
        toolsCount: { type: integer }
    BundleDetail:
      type: object
      properties:
        name: { type: string }
        description: { type: [string, "null"] }
        enabled: { type: boolean }
        createdAt: { type: integer, description: Unix ms epoch }
        updatedAt: { type: integer, description: Unix ms epoch }
        tools: { type: array, items: { $ref: "#/components/schemas/BundleToolRef" } }
        composites:
          {
            type: array,
            items: { type: string },
            description: "Names of composite (macro) tools this bundle also exposes.",
          }
    InstallLink:
      type: object
      description: >
        Metadata for a bundle install link. The raw token is never included in
        list/detail responses — only a short tokenPrefix for identification.
      properties:
        id: { type: integer }
        bundleName: { type: string }
        tokenPrefix: { type: string, description: "First characters of the token, for display only." }
        mcpKeyId: { type: integer, description: The scoped MCP API key auto-provisioned for this link. }
        createdBy: { type: [string, "null"] }
        createdAt: { type: integer, description: Unix ms epoch. }
        expiresAt: { type: [integer, "null"], description: "Unix ms epoch, or null for no expiry." }
        revokedAt: { type: [integer, "null"], description: "Unix ms epoch when revoked, or null if active." }
        lastUsedAt: { type: [integer, "null"], description: "Unix ms epoch of last successful resolve, or null." }
...
    ApiError:
      type: object
      description: >
        Standard error envelope returned by the bridge's REST endpoints (not
        MCP transport endpoints, which use JSON-RPC 2.0 error objects instead).
      required:
        - error
      properties:
        error:
          type: object
          description: Container for the structured error detail.
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: >
                Machine-readable error code in SCREAMING_SNAKE_CASE.
                Known values: VALIDATION_ERROR, CLIENT_NOT_FOUND, DISCOVERY_ERROR,
                UNAUTHORIZED, FORBIDDEN, RATE_LIMITED.
              example: VALIDATION_ERROR
            message:
              type: string
              description: >
                Human-readable description of what went wrong, suitable for
                displaying in a developer console or log.
              example: "Missing required fields: name, health_url"
            retry_after:
              type: integer
              description: >
                Seconds the caller should wait before retrying. Present only on
                a 429 RATE_LIMITED response (also echoed in the Retry-After
                header); omitted on every other error.
              example: 45
            request_id:
              type: [string, "null"]
              description: >
                Correlation id for this request (from res.locals.requestId),
                echoed back so a caller can quote it when reporting an issue.
                Present on almost every error response in the admin API;
                absent (or omitted) on the handful of handlers that don't
                thread it through (notably GET /clients and GET
                /clients/{name}/tools's error bodies, and /metrics's disabled
                response).
              example: "3f9c9a2e-9c34-4b3a-9a0e-4f7a6f6c9a1b"
      example:
        error:
          code: VALIDATION_ERROR
          message: "Missing required fields: name, health_url"
          request_id: "3f9c9a2e-9c34-4b3a-9a0e-4f7a6f6c9a1b"

    # -------------------------------------------------------------------------
    # JsonRpcMessage  (internal — MCP transport)
    # -------------------------------------------------------------------------
...